# docs/swagger.yaml
basePath: /api
definitions:
  handler.AsyncUpdateRequestRatioRequest:
    properties:
      Ratio:
        type: number
      key:
        type: string
    required:
    - Ratio
    - key
    type: object
  handler.CostRequestDetailResponse:
    properties:
      Max_volume:
        type: integer
      Min_volume:
        type: integer
      Ratio:
        type: number
      created_at:
        type: string
      id:
        type: integer
      price_request_to_costs:
        items:
          $ref: '#/definitions/handler.PriceRequestToCostDetailResponse'
        type: array
      status:
        type: integer
    type: object
  handler.CostRequestInfoResponse:
    properties:
      item_count:
        type: integer
      request_id:
        type: integer
    type: object
...
  handler.CostsRequestsFilterResponse:
    properties:
      ClosedAt:
        type: string
      CreatedAt:
        type: string
      FormedAt:
        type: string
      Max_volume:
        type: integer
      Min_volume:
        type: integer
      ModeratorID:
        type: integer
      Ratio:
        type: number
      Status:
        type: integer
      UserID:
        type: integer
      id:
        type: integer
      username:
        type: string
    required:
    - username
    type: object
...
  handler.PriceRequestToCostDetailResponse:
    properties:
      cost_id:
        type: integer
      cost_price:
        type: number
      cost_title:
        type: string
      img:
        type: string
    type: object
...
  handler.UpdateCostRequestResponse:
    properties:
      Max_volume:
        type: integer
      Min_volume:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@DIA.example.com
    name: API Support
    url: https://github.com/your-org/DIA_Backend
  description: This is the backend API for DIA (Information and Analytical Data) system.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: DIA Backend API
  version: "1.0"
paths:
...
  /cost-requests:
    get:
      consumes:
      - application/json
      description: Get a list of cost requests with optional filtering
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: integer
      - description: Filter by date from (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: Filter by date to (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.CostsRequestsFilterResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get cost requests
      tags:
      - cost-requests
  /cost-requests/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a cost request
      parameters:
      - description: Cost Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete cost request
      tags:
      - cost-requests
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific cost request
      parameters:
      - description: Cost Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CostRequestDetailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get cost request by ID
      tags:
      - cost-requests
    put:
      consumes:
      - application/json
      description: Update an existing cost request
      parameters:
      - description: Cost Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cost request update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateCostRequestResponse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update cost request
      tags:
      - cost-requests
  /cost-requests/{id}/async-update:
    put:
      consumes:
      - application/json
      description: Update ratio in request asynchronously using secret key
      parameters:
      - description: Cost Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update ratio
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.AsyncUpdateRequestRatioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Asynchronously update ratio in request
      tags:
      - cost-requests
  /cost-requests/{id}/form:
    put:
      consumes:
      - application/json
      description: Form a draft cost request into a submitted request
      parameters:
      - description: Cost Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Form cost request
      tags:
      - cost-requests
  /cost-requests/{id}/reject:
    put:
      consumes:
      - application/json
      description: Reject a cost request (moderator action)
      parameters:
      - description: Cost Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Reject cost request
      tags:
      - cost-requests
  /cost-requests/{id}/resolve:
    put:
      consumes:
      - application/json
      description: Resolve a cost request (moderator action)
      parameters:
      - description: Cost Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Resolve cost request
      tags:
      - cost-requests
  /cost-requests/costRequestInfo:
    get:
      consumes:
      - application/json
      description: Get information about current user's draft request
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CostRequestInfoResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get draft request info
      tags:
      - cost-requests
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"